requirements.txt
dependency-injector
fastapi
uvicorn
requests
lxml
beautifulsoup4
shikimori_api
pygelbooru
httpx
blurhash-python
aioredis<2  # TODO: Update example to work with aioredis >= 2.0

# For testing:
pytest
pytest-asyncio
pytest-cov

